requirements.txt
llama-index-core==0.11.9
openai==1.45.0
python-dotenv==1.0.1
llama-index-readers-file==0.2.1
PyMuPDF==1.24.10
